[metasrv] refactor: move kv api integration test to metasrv/tests/flight/metasrv_flight_kv_api.rs
